+++ src/form/select.rs
@@ -27,6 +27,9 @@ pub struct SelectProps {
     /// Display a loading spinner within this component.
     #[prop_or_default]
     pub loading: bool,
+    /// Disable this component.
+    #[prop_or_default]
+    pub disabled: bool,
 }
 
 /// A wrapper around an HTML `select` tag.
@@ -78,6 +81,7 @@ impl Component for Select {
                 <select
                     name=self.props.name.clone()
                     value=self.props.value.clone()
+                    disabled=self.props.disabled
                     onchange=self.link.callback(|change: ChangeData| match change {
                         ChangeData::Select(data) => data.value(),
                         _ => unreachable!("invariant violation: received non-select change event from a select element"),
@@ -117,6 +121,9 @@ pub struct MultiSelectProps {
     /// Display a loading spinner within this component.
     #[prop_or_default]
     pub loading: bool,
+    /// Disable this component.
+    #[prop_or_default]
+    pub disabled: bool,
 }
 
 /// A wrapper around an HTML `select` tag with the `multiple=true` attribute.
@@ -170,6 +177,7 @@ impl Component for MultiSelect {
                     size=self.props.list_size
                     name=self.props.name.clone()
                     value=self.props.value.join(",")
+                    disabled=self.props.disabled
                     onchange=self.link.callback(|change: ChangeData| match change {
                         ChangeData::Select(data) => {
                             let opts = data.selected_options();